Simplifies user experience by always having a Signal have the latest data
We now guarantee the newest data to be sent to a Subscriber. This
removes the need to check versions after a loop.  This simplifies
client code.

It complicates this code by introducing a signaler that throws away
stale entries when a newer entry comes in.  To the user they just see
the channel, but behind the scenes, lots of stuff going on.