Optimize CleanMessages Middleware
Because the new messages are always appended to the end, when we find
the first message that doesn’t need to be deleted, we just need to stop
the loop and copy the array from that index.

If i == 0, then we don’t need to make a copy at all.