Fix problem where data scrolled at wrong point
I hate magic, but it seems like “3” is the magic number to check the
len(data) against the Height to see if we are going to scroll off,
regardless of font size.

The terminal program does not list what width/height are measured in
anyways.

Also, now handles resizing text if window changes size.