Fix for issue #3
Greatly simplified the signaler class.  It was using channel semantics,
when it really is easier to do this with a lock and mutex.

The bug was mainly held up in using go s.cast().  This didn’t allow for
keeping order.  We could have solved this by having cast ignore any
update that didn’t have a higher version number.

But it is simpler and in the common case faster to just make this
synchronous.  It also simplifies the code with removal of the WaitGroup.

Tests were updates to run in parallel and have a better test for the
signaler object.